Per thread custom main loops
Some tasks are difficult to implement having only events. Having a
dedicated loop doing the processing is much better. This modification
offers the possibility to replace the simple per thread main loop in
event_processor with a custom event loop implemented in a class derived
from event_receiver.

The intention is to use this to implement a demultiplexer with FFmpeg.
Whenever the demultiplexer needs to read new data it calls the method
event_processor::dequeue_and_process() to process queued events. One of
the events should provide new data.