Decoding and Displaying Video
FFmpeg is used to decode the video stream.
The X11 extension Xv is used to display the video frames.

VideoDecoder: Receives VideoPacketEvent events from the Demuxer and
              empty  events from the VideoOutput. These events
              are queued in packetQueue and frameQueue.
              Filled XFVideoImage events with YUV images are sent back
              to VideoOutput.

The XFVideoImage contains a shared memory buffer used to display the buffer
with XV. When a buffers has the wrong size a ResizeVideoOutputReq event
is sent to VideoOutput. VideoOutput resizes the window and allocates a new
buffer when receiving the ResizeVideoOutputReq event.

VideoOutput: Displays the received XFVideoImage frames using PTS and
             ShowNextFrame timer events.

XlibFacade: Interface classes for X11 and the Xv extension.

XlibHelpers: Debug functions printing X11 structures.

timer.hpp: Added operators for timespec_t (+, -, <, <<)
           Conversion between timespec_t and double