Introduced AudioFlushedInd to avoid hanging video after seeking
When seeking the AudioOutput and the VideoOutput are flushed independently.
AudioOutput may send a AudioSyncInfo to VideoOutput, that is transmitted
before the flush of AudioOutput and received after the flush of the
VideoOutput. In this case the frameTimer calculation is not correct.
Jumping by 10 seconds forwards results in a 10 second timer.

Solution:
- AudioOutput sends a AudioFlushedInd to VideoOutput after flushing its
  buffers.
- In VideoOutput an ignoreAudiSync counter is introduced. The counter is
  incremented when receiving a FlushReq, it is decremented when receiving
  a AudioFlushedInd.
- AudioSyncInfo events are only accepted when ignoreAudioSync is 0.