Fixed page size of adjustments
The value of an adjustment is in the range [lower, upper - page_size].

Having the page size of the position adjustment set to 60, does not allow
to seek to positions within the last 60 seconds.

Set position adjustment page size to 1. This avoids immediately sipping
to the next file when moving the position scollbar to its end.

Volume adjustment position is always set to 0. This allows to adjust the
complete range.