Playing audio only and video only files
Playing media files having only audio or only video. Playing video stream
only, when audio stream is not supported by the decoder and vice versa.

VideoOutput is notified with NoAudioStream when no audio stream is
available. This is necessary, since VideoOutput must not wait for
AudioSyncInfo. By simulating the AudioSyncInfo the differences are
minimized.

AudioOutput is notified with NoVideoStream when no video stream is
available. This is necessary, since AudioOutput has to provide time
information for the GUI in this case. Currently AudioOutput sends
AudioSyncInfo events to the MediaPlayer in this case. This is done, since
the message is not send on second boundaries. Using the AudioSyncInfo
instead of NotificationCurrentTime allows the GUI to do some fine tuning.
With MP3 files I have checked this seems not to be necessary.

Demuxer: Removed SystemStreamFailed state. It is now implemented as
substate of SystemStreamOpened using the boolean variable
systemStreamFailed.

OpenFile and CloseFile procedures between MediaPlayer and Demuxer are now
request-response/failure procedures. If opening the system stream failed
or when opening both the audio and the video stream failed, then the
Demuxer responds with OpenFileFail. Everything is closed by the demuxer
in this case. Otherwise the Demuxer responds with OpenFileResp. This
especially includes the case when opening succeeded for one stream and
failed for the other stream.

MediaPlayer skips to the next file in the playlist when opening the file
failed.

Demuxer sends an OpenAudioStreamFailed/OpenVideoStreamFailed to
MediaPlayer before sending the OpenFileResp, when the system stream
contained an audio/video stream, but the AudioDecoder/VideoDecoder
failed to open it. The process method is virtual, this allows the GUI
class to add an error message.

Demuxer: Seek by default uses the video stream, when no video stream is
available the audio stream is used. After seeking the stream may no longer
be in the failed state. This is since reading EOF results in setting the
failed state.

In case of any system stream read frame error the "end of stream"
procedure is started by sending EndOfSystemStream, EndOfAudioStream and
EndOfVideoStream.