Pause implementation for audio devices not supporting pause.
Not all ALSA devices do support snd_pcm_pause. In this case snd_pcm_drop
is used to immediately stop playback. When playback restarts the buffer
in the audio device is refilled with the dropped samples.

To support this the frame queue is replaced with a list and a currentFrame
pointer/iterator. New frames received from the AudioDecoder are added at
the end of the list. The frames at the front of the list are already
copied into the buffer of the audio device. They are only kept to restore
them when restarting playback after pause.

currentFrame points to the frame samples are read from currently and are
added to the end of the buffer of the audio device.

The number of additionally needed AFAudioFrame objects for this feature
depends on the decoder. The variables numAudioFrames and numPostBuffered
are introduced to allocate these objects as needed.