Integrated X11 Xvideo output into MainWindow
GtkmmMediaPlayer is derived from MediaPlayer and Gtk::DrawingArea.
It is used as widget in MainWindow.

All X11 events are received via gtkmm. In XlibFacade no X11 events are
polled anymore. The relevant events are now sent via GtkmmMediaPlayer and
VideoOutput to XlibFacade, i.e. to class XFVideo.

GtkmmMediaPlayer::on_realize() is called after the widget is desplayed.
When this method is called it is possible to determine the window id.
display and window id are sent to XFVideo using a WindowRealizeEvent.

GtkmmMediaPlayer::on_configure_event(...) sends a WindowConfigureEvent
indicating a resized window.

GtkmmMediaPlayer::on_expose_event(...) sends a WindowExposeEvent
requesting a window update.

GtkmmMediaPlayer provides the signal notificationResizeOutput. This signal
is emitted when the video size has changed. It is used to adapt the size
of MainWindow.

The XVideo output is integrated into MainWindow. This requires to run
VideoOutput and the GUI within the same thread.

VideoOutput creates the XFVideo object when the first WindowRealizeEvent
is received. Until this event the window id is not available. The XFVideo
object is not deleted when a CloseVideoOutputReq is received, a black
frame is displayed.

MediaPlayer::open() must not be called before the Xv widget is realized.
Calling open method in GtkmmMediaPlayer::on_realize() and not in main.

Removed classes XFDisplay and XFWindow. gtkmm toolkit is used for this.