Fullscreen mode
Visibility of toolbar, menubar and statusbar is handled separately for
window mode and fullscreen mode. By default all three are shown in window
mode and none is shown in fullscreen mode. The visibility of these widgets
is also modified when the window manager is used to enter or leave
fullscreen mode.