Clipping with mouse
The GtkmmMediaPlayer widget captures right mouse button clicks. The
shape of the cursor changes depending on the position within the widget.
When the right mouse button is clicked a part of the edge is clipped
depending on the mouse position. When clicked in the middle of the
window any clipping is removed. A ClipVideoEvent with window coordinates
is send via VideoOutput to XlibFacade. XlibFacade converts to them into
video coordinates and calculates the new clipping and destination area.

With my ATI/AMD graphic card I faced the problem that XvShmPutImage
doesn't use the source area as expected. For that reason I implemented
a clip method in XFVideo.

Bug fixes:

Now using offsets and pitches arrays of XvImage to get correct results
for any clipping area. Also implemented two methods creating test patterns
for debugging purposes.

Fixed memory leak in XFVideoImage. XShmDetach wasn't called.