Using correct zoom factor for clipped video
For clipped video the window size calculation is fixed to get the requested
zoom factor. After clipping the window is resized to retain the existing
zoom factor. Zoom factor is displayed in the statusbar. round() is used
before converting from double to int. zoom is rounded to multiples of 1%.
This avoids that the value drifts away.

MainWindow:
- Receives NotificationVideoSize from XFVideo. It contains video size,
  window size, size of clipped area in unzoomed video and in zoomed video,
  i.e. video size visible in window.
- ResizeVideoOutputReq is now sent from VideoDecoder to VideoOutput only.
  It was also sent to GUI.