Hot keys for main player functions
The supported hot keys are:
- Cursor left: Jump 10 seconds back
- Cursor right: Jump 10 seconds forth
- Space: Play/Pause
- Cursor Up: Increase volume
- Cursor Down: Decrease volume
- m: Toggle mute
- Page Up: Jump to previous title
- Page Down: Jump to next title
- Esc: Terminate application

The delta used by the jump and volume actions is multiplied with 3 when
the shift key is pressed. It is multiplied with 6 when the control key is
pressed.

GtkmmMediaPlayer grabs the focus to get key press events.

Either action play or action pause is shown.