ChannelConfigWindow is using and updating config file
Parsing and generating the configuration file is not done in the GUI
thread.

ConfigFile: event_receiver class running in an own thread.
  After receiving CommonInitEvent it emits  containing
  the station list read from the configuration file.
  When receiving ConfigurationData the configuration file is updated.

Configuration file: ./playerrc, when it exists. Otherwise ~/.playerrc.
  If no file exists yet ~/.playerrc is created. If environment variable
  HOME is not defined ./playerrc is created.

ChannelConfigWindow: Emits signalConfigurationDataChanged signals to
  update the configuration file. The dont_save flag is used to avoid
  unnecessary file updates when subsequent modifications are done.

TemporaryValue, TemporaryEnable, TemporaryDisable: Classes to temporarily
set a variable. Objects of these classes are created on the stack. When
the object is created a referenced variable is set to a new value, when
the object is deleted the referenced variable is set back to the original
value.