File open dialog
A modal file open dialog is shown. Multiple files can be selected. When
opening the files the current play list is replaced with the selected
files. Playback starts on the new list.

The PlayList class is used by the SignalDispatcher class and by the
MediaPlayer class. Since both classes are running in the GUI thread,
no mutex is needed when accessing PlayList.

PlayList is allocated on stack and passed by reference.