Channels menu added
Added a Channels menu to menu bar and popup menu. It contains:
- Channel Config Window: ToggleAction, opening/closing the window to
                         configure channels.
- Previous: Action, switches to previous channel
- Next: Action, switches to next channel
- A dynamically adapted list of RadioActions for all configured channels.
  Selecting an entry switches to that channel. The selected entry is
  updated when a new channel is tuned.

SignalDispatcher::on_channel_selected(int) is called when the user
selects a channel using the menu and also when the application updates
the radio button group due to a received on_tuner_channel_tuned event.
Only in the first case a signalSetFrequency is emitted.
m_isEnabled_signalSetFrequency is used to distinguish between both cases.