Implemented menu File-Close
Closing a file stops playing that file and removes the file from the play
list. If it wasn't the last file in the play list, then the next file is
automatically played. Otherwise, i.e. if it was the last file, then
playback stops and the first file in the play list is selected.
Subsequently pressing play again starts the play list at the beginning.

When MediaPlayer finished playing the last file without using the Close
menu, then the behaviour is the same. Playback stops and the first file
of the play list is selected.