Active channel radio action is always up to date
The Channels menu contains a list of radio buttons. The active member
indicates the tuned channel. At program start up TunerFacade reads the
currently set frequency from the tuner and sends an event to the
SignalDispatcher to activate the correct radio button.

If a frequency is tuned, that is not contained in the channel list, then
none of the (visible) radio buttons is activated.

At program start up the SignalDispatcher gets the ConfigurationData event
from ConfigFile and the ChannelData event from TunerFacade. Both senders
are executed in different threads. For that reason the events may arrive
in any sequence. When the ChannelData event is received before the
ConfigurationData event, then processing of the ChannelData event is
deferred until the channel list is available.

The first ChannelData event sent by the TunerFacade contains the frequency
only, but no standard name and no channel name. ChannelConfigWindow::
on_tuner_signal_detected ignores this event to avoid adding a new entry
to the channel list.