Fixed inconsistent startup behaviour.
Sometimes TV playback started automatically on startup, sometimes not.

The problem is that the sequence in which the SignalDispatcher receives
the configuration and the currently tuned channel depends on how the
threads are scheduled.

When the channel configuration was received first
(SignalDispatcher::on_configuration_data_changed) and subsequently
the currently tuned frequency (SignalDispatcher::on_tuner_channel_tuned),
then playback didn't started automatically. If the 2 events were received
in reverse order then it started automatically.

Playback is actually started in SignalDispatcher::on_channel_selected
when the flag m_isEnabled_signalSetFrequency is true.

In SignalDispatcher::on_tuner_channel_tuned the flag is temporarily set
to false, this is mainly done to avoid recursion. Setting the RadioAction
results in calling SignalDispatcher::on_channel_selected. Here it is
necessary to suppress sending signalSetFrequency back to TunerFacade to
avoid a recursion. The flag also suppresses autostart of playback.

The fix now is to also set the flag temporarily to false in
SignalDispatcher::on_configuration_data_changed when the deferred setting
of the correct RadioAction for the tuned channel is necessary.

By doing this TV playback will never start automatically. That's the best
solution without storing the state of the last program session.