Added PlayListWindow
Added a window that shows the content of the play list. The menu
"View/Play List Window" opens that window.

Open points:
- Current title isn't highlighted.
- Inefficient implementation.
- No interaction possible with play list window.

Related classes are:
PlayListWindow: This class implements the window with a Gtk::TreeView
   showing the play list and a context menu.

PlayListTreeModel: This class implements a Gtk::TreeModel that is used by
   tree view to access the play list.

PlayList: This class already existed. It implements the play list
   interface used by the MediaPlayer class which has no dependencies
   to gtkmm. Now this is a vitual base class. The class is still using
   a list. This isn't really efficient anymore, since most play list
   accesses are done by index from the PlayListWindow.

GtkmmPlayList: This class is derived from PlayList. It adapts the
   PlayList class to gtkmm.