"Remove Entry" and "Clear All Entries" in PlayListWindow context menu
"Remove Entry" removes the title selected in the tree view from the play
list. If the removed entry is not the currently played title, then this
is trivial. If it is the currently played title, then signal_file_close
is emitted. This signal is processed by the SignalDispatcher. It triggers
the same behaviour as using the File/Close menu. (Closing file;
selecting next file; start playback again, if it was playing)

"Clear All Entries" removes all entries from the play list. This always
includes the currently played file. That file is closed by emitting the
Signal_close. This signal directly goes to the MediaPlayer. The more
complex SignalDispatcher behaviour isn't necessary here, since there is
no next file.