Drag and drop files into the PlayListWindow
Drag and drop a single file or a list of files into the PlayListWindow
adds the files to the play list at the position where they are dropped.
No events are sent to MediaPlayer, i.e. current file is not changed.

This works with Thunar, the Xfce file manager.

With konqueror, the KDE file manager, it does not work.