Reanimated audio video sync test application
synctest is a test application to see, if audio and video are in sync.
It uses SyncTest, AudioOutput, VideoOutput and the sub-classes to access
the output devices. Everything else is stubbed out. The trick is
basically done in player/GeneralEvents.hpp, which is included first in
all files linked to synctest.

Integrated the synctest build into the autotools build.