Recorder updates duration displayed in GUI
The Recorder permanently determines the current duration of the recording
and updates the GUI by sending NotificationFileInfo events.

The GtkmmMediaRecorder signal notificationDuration is connected to the
same objects as the corresponding signal emitted by GtkmmMediaPlayer.

FFmpeg is used to determine the current duration of the recording. For
reading the temporary file Recorder uses StorageProtocol. This is the
base class of PvrProtocol. It just reads the temporary file and does not
start a recording.

Note: Currently the StorageProtocol class is not needed, since the
      recording is still stored in a normal mpeg file. In the future this
      will be change to be able to remove old time shifting fragments.