Inhibiting screen saver via dbus SimulateUserActivity method
The implementation is tested with the KDE and Gnome screen savers. The
MediaPlayer notificationCurrentTime signal is connected to InhibitScreenSaver.
When that signal is received during playback a SimulateUserActivity request
is sent via dbus to the screen saver. The used KDE screen saver dbus object
is (org.freedesktop.ScreenSaver, /ScreenSaver) the Gnome object is
(org.gnome.ScreenSaver, /).

Using dbus-cxx as libdbus binding.

gcc needs the option  to compile the dbus-cxx headers. The
future/promise code used in PvrProtocol does not compile with -std=c++0x.
Only compiling InhibitScreenSaver.cpp with that option. With autotools
this is only possible by introducing a new library.