Menu File/Quit immediately stops playback.
Implemented Demuxer open/close file procedures for all system stream states.

The menu File/Quit just terminated the Gtk main loop. Especially audio
playback didn't stopped immediately with that implementation.

Now the SignalDispatcher sends a CloseFileReq to Demuxer and waits for the
reponse before terminating the Gtk main loop. This way audio immediately
stops.

Demuxer: Now Demuxer always sends a response for each OpenFileReq and
each CloseFileReq message. When a CloseFileReq is received in state Closed
then another CloseFileResp is sent. When a OpenFileReq is received in state
Opened then a OpenFileFail is sent. OpenFileReq and CloseFileReq are both
deferred when another procedure is still in progress, i.e. in the states
SystemStreamClosing and SystemStreamOpening.

In main SignalDispatcher::signal_close was connected twice to
GtkmmMediaPlayer::close.