Always delay termination until MediaPlayer close procedure is terminated
Now program termination is also delayed when the window close button
provided by the window manager is used to terminate the application.

The Gtk main loop is started by calling Gtk::Main::run without window
parameter. In this case it is necessary to explicitly call Gtk::Main:quit()
to terminate the main loop. This call is done when the MediaPlayer close
procedure is terminated. The MainWindow is hidden when the procedure
starts. This gives a quick user feedback.