Fixed calculation of EOS timer in AudioOutput
The problem occurred when using the virtual Alsa device of PulseAudio.
For this device the latency never goes down to zero and the EOS timer
is restarted infinitely.

Now the number of samples still in the playback buffer is used to
calculate the EOS timer.

Note: This method does not take the additional hardware delay into account.