Automatically skipping to next file for audio or video only media
MediaPlayer class didn't skipped to the next file in the play list when
an audio or video only file was played.

When opening a file the MediaPlayer class is notified by the Demuxer if
the stream has no audio or no video. Now this information is stored and
used when receiving the EndOfAudioStream and EndOfAudioStream events.