Audio/Video Sync and SyncTest
From class AudioOutput AudioSyncInfo events are sent to class VideoOutput.
The event contains the current time and the PTS of the audio frame
currently processed by the DAC. VideoOutput uses this information to
calculate when to display the next Video frame and starts the timer.

SyncTest: When compiled with macro SYNCTEST defined a test application
is built. It periodically plays the scale of C major (C-Dur-Tonleiter)
with one second and a unique picture for each tone (from dark to light).