Sending title and time from MediaPlayer threads to GUI
ControlWindow displays current title and current time. Title is sent from
Demuxer and time is sent from VideoOutput to gtkmm ControlWindow via
GtkmMediaPlayer. GtkmMediaPlayer is the only component having dependencies
to the player directory and to gtkmm. All files in the player directory
have no dependencies to gtkmm. The GtkmmMediaPlayer class is derived from
MediaPlayer and is placed in the gui directory.

The MediaPlayer class now also derives from event_receiver. This allows
to use the same interface to send events to class MediaPlayer. The
MediaPlayer class processes these events in the GUI thread. When an
event is queued for the MediaPlayer the concurrent_queue template class
creates on object of class MediaPlayerThreadNotification. This results in
notifying gtkmm to call the MediaPlayer::processEventQueue method by using
a Glib::Dispatcher and connecting its signal to the MediaPlayer method.

MediaPlayer::processEventQueue() processes all queued events by using the
event_processor class. This results in calling the virtual process
methods implemented in GtkmmMediaPlayer. In these process methods sigc++
signals are emitted as also done by other gtkmm widgets. These signals
are connected to ControlWindow slots to display current title and time.