Deinterlacer support for close, flush and end of stream procedures
VideDecoder sends the events CloseVideoStreamReq, FlushReq and
EndOfVideoStream via Deinterlacer to VideoOutput. This is done for both
interlaced and non-interlaced video. It is not necessary to distinguish
between these two cases here. The response messages are still send
directly from VideoOutput to Deinterlacer.

CloseVideoOutputReq: Video is closed. All XFVideoImage objects are deleted.
Member data of Deinterlacer is set back to the default. VideoDecoder must
do the same with its m_topFieldFirst.

FlushReq: This procedure is used when seeking to another file position.
In this case XFVideoImage objects are not deleted. The purpose of flush is
to start as quick as possible at the new position. Everything is removed
from the imcoming queues. XFVideoImage objects are sent back to VideoDecoder.
This is the same as already done by VideoOutput.

EndOfVideoStream: The Deinterlacer receives this indication after it has
received the last frame of the video stream. It forwards the indication to
VideoOutput after sending the last frame. The deinterlacer needs up to 4
subsequent frames to calculate the deinterlaced frame. For the last frames
this if of course not possible and these frames are sent interlaced to
VideoOutput. The empty frames are sent back to VideoDecoder. It is
important to not delete XFVideoImage objects here, since it is still
possible that the seek function is used to jump to another file position
until the last frame is displayed.

For the open procedure nothing has to be done by Deinterlacer.