Player and GUI configuration is stored in resource file
Added a new tab PlayerConfigWidget to ConfigWindow.

Extended ConfigurationData structure with substructures for GUI and player
configuration. In ConfigFile the boost::spirit parser stores these new
structures in the resource file:

gui
{
    fullscreen = false;
    window
    {
        menuBar = true;
        toolBar = true;
        statusBar = true;
    };
    fullscreen
    {
        menuBar = false;
        toolBar = false;
        statusBar = true;
    };
};

player
{
    useOptimalPixelFormat = true;
    useXvClipping = true;
    deinterlacer = "Blur: Temporal";
};

The boost::spirit parser and generator are using attributes.

On the signal-slot interface between MediaCommon/ConfigFile and the GUI
objects PlayerConfigWidget, SignalDispatcher, ChannelConfigWidget a
boost::shared_ptr<ConfigurationData> is used. When parsing ConfigFile,
running in an own thread, generates that object and sends it to the GUI
thread. Now the GUI thread owns the object. After modifying
ConfigurationData in the GUI thread the data, not the object, is sent back
to ConfigFile. MediaCommon::saveConfigurationData makes a copy! This
function is running in the GUI thread. This implementation avoids that
ConfigurationData is simultaneously used by 2 different threads.

All GUI objects are always working on the same ConfigurationData object.
This is no problem, since all GUI objects are executed in the same thread.

NotificationDeinterlacerList: Event sent by Deinterlacer to let the GUI
  know which deinterlacers are available and which one currently is used.

SelectDeinterlacer: Event sent from GUI to Deinterlacer to select another
  deinterlacer.

EnableOptimalPixelFormat/DisableOptimalPixelFormat: Events to configure the
  VideoDecoder.

EnableXvClipping/DisableXvClipping: Events to configure XlibFacade.