Using prefix TRACE for tracing macros.
DEBUG and ERROR are not good names for tracing macros. Especially DEBUG
is used by system header files to select the debug mode. As a result it
wasn't possible to include header files in arbitrary order.

Now including own header files first and than the system headers. This
is better to find missing includes.