Do not skip forward when file is already opened
For simplicity the GUI always sends a open and a play event when the play
button is pressed. By doing this the Demuxer also gets a OpenFileReq when
play is pressed in pause mode. In this case the file is already opened and
the MediaPlayer class should not skip to the next file.