Integrating deinterlacer code into the GNU Auto Tools build
Replaced ARCH_X86 and ARCH_X86_64 with the macros ARCH_X86 and
ARCH_X86 predefined by compiler.

In the *.inc files the IS_MMX, IS_SSE and IS_3DNOW are kept. These files
are included for multiple times into the same source file with different
IS_* settings. In other files the macros are replaced with __MMX__,
__SSE__ and __3DNOW__.