Removed code to inhibit KDE and Gnome screen savers
Screen savers were disabled via DBus. The dbus-cxx C++ wrapper is used to
access the DBus library. The Dispatcher::watch_thread_main() function of
dbus-cxx is running in an thread created by the wrapper library. It throws
an exception when select() returns -1. I don't see any option to catch this
exception.

As a result the application sporadically was interrupted.

Another reason for removing is, that dbus-cxx is not part of Ubuntu.