Introduced std::unique_ptr<XFVideoImage> for reliable X11 resource handling
Accessing the X11 library is allowed from a single thread only. Using the
std::unique_ptr smart pointer ensures that XFVideoImage and DeleteXFVideoImage
objects are deleted in the context of the VideoOutput object only, i.e. in
the X11/GUI thread.

The former implementation using boost::shared_ptr did not ensure this. Due
to process scheduling the smart pointer in an event sender may be deleted
after the smart pointer in the event_receiver. In case of the
DeleteXFVideoImage event this results in calling the destructor of the
XFVideoImage object in the context of the VideoDecoder. This sporadically
resulted in X11 protocol errors, a blocking application and even crashes.

unique_ptr<T> destructor needs the complete type definition of T. Moved
constructor and destructor implementation of several classes from hpp to
cpp file.

Using raw pointers at several private interfaces, where the ownership
should not be passed to the called function.

Implemented VideoOutput::frameQueue as std::list. This is needed in
VideoOutput::startFrameTimer to read the PTS from the next frame in the
frameQueue without removing the element from the queue and without
transferring the ownership.