Creating one Server instance for each accepted TCP connection
In the server process a new tcp_connection and a new Server object is
created for each accepted TCP connection. When the connection is closed
both objects are deleted again in a controlled way after all queued
messages are processed.

event_receiver/event_processor do not store a shared pointer to the
receiver into the queue. So potentially the event_receiver exists when
putting an event into the queue and it is deleted when removing the event
from the queue. The shared pointer isn't put into the queue in general to
avoid unnecessary overhead.

As a solution the ConnectionReleasedIndication/ConnectionReleasedConfirm
procedure is used when the TCP connection is terminated by the remote side.
The ConnectionReleaseRequest/ConnectionReleaseResponse procedure is used
when the connection is closed by the process itself. The first event
of both procedures contains shared pointers to both the tcp_connection
object and the Receiver object. The second message of both procedures only
contains a shared pointer to the object the message is sent to. After an
object sends one of the four messages it is not allowed to send further
messages to its peer, with one exception. It is still allowed to send the
second message of the other procedure. This is needed to support the case,
that both sides simultaneously close the connection.

As a side effect the Server object is deleted in the thread of its
event_processor.

The client terminates itself after receiving the response from the server.
Boost Asio gets no next procedure and io_service.run() returns as a result.
Then the Appl destructor terminates the client thread and the whole process
terminates.