tcp_connection interface consistency: queue_event method
The class tcp_connection now has overloads of the queue_event method to
accept control events for the TCP connection and it also has overloads
to accept events forwarded over the TCP connection.

queue_event is the same interface as event_receiver. I.e. the code is the
same for sending an event to an event_receiver or to a tcp_connection
proxy.