sinemad process is started in the background
In future sinemad will implement everything needed to automatically record
TV programs in the background. With this modification only the TunerFacade
is moved to sinemad. The classes MediaReceiver and GtkmmMediaReceiver are
removed.

DaemonProxy and GtkmmDaemonProxy are running in the sinema process. They
implement the interface to access sinemad:
- Starting the process with the help of process_starter.
- Establishing a TCP connection with the help of tcp_connector and
  tcp_client.
- Sending messages to and receiving messages from sinemad with the help of
  tcp_connection.

Daemon:
- This class is using tcp_acceptor to setup a TCP server.
- For each accepted connection a Server instance is created.
- One object instance of this class exists in sinemad.

Server:
- One instance is created for each accepted TCP connection.
- Acts as interface between tcp_connection and TunerFacade.

SinemadInterface:
- Defines events exchanged between sinema and sinemad.