Fixed zoom and window resize mechanism
The old implementation occasionally calculated a wrong window size in
MainWindow::zoom. The problem was an inconsistency between the MainWindow
size and the GtkmmMediaPlayer widget size. It seems that the zoom function
was executed when only one of the two objects was updated.

The implementation supports:
- Resetting the zoom factor to 100% when a new window is started.
- Resizing the window using the window manager: The zooming factor is
  changed.
- Changing zoom factor via menu: The window size is adapted to the new
  zoom factor.
- Enabling and disabling menu/tool/status bar: The size of the video
  widget stays unaltered, but the window size is changed.
- Clipping the video using the mouse: The zoom factor stays constant.
  Video widget and window size are adapted.
- Entering and leaving fullscreen mode: After leaving the fullscreen
  mode the old zoom factor/window size is restored.

GtkmmMediaPlayer::zoom(double): This function is called when the zoom
factor is changed via a menu of the application itself.
GtkmmMediaPlayer::dontZoom(): This function is called when the window
size has to be adapted, but zooming should stay constant.

In both cases the main window may have to be resized. This is done by
temporarily setting the size request (minimum widget size) of the
GtkmmMediaPlayer widget to the expected size. Then the application tries
to set the size of the main window to a minimum (1,1). This starts the
size requisition phase. The main window recursivly asks all its subwidgets
for their minimum size to calculate the needed window size. An idle
callback is attached to clear the size request again. This again allows
to shrink the window using the window manager.

Finally GtkmmMediaPlayer::on_size_allocate is called with the new widget
size. Since the mechanism was initiated by the application itself nothing
has to be done in this case.

GtkmmMediaPlayer::on_size_allocate is also called when the window is
resized using the window manager. In this case the zoom factor has to
be updated.

MainWindow: Moved most zoom/resize related code to GtkmmMediaPlayer.
            The new slot resizeMainWindow starts the size requisition.