Deinterlace visible lines only
VideoOutput notifies Deinterlacer about the current clipping area.
Deinterlacer uses this information to deinterlace displayed lines only.

In XFVideo::resize now also a NotificationClipping event is sent. This
function is called when a new video starts and clipping is reset. The
event is needed to update the GUI (wasn't done) and the deinterlacer
(new feature).