Do not automatically skip to next file when displaying photos
With this modification it is necessary to manually skip to the next file
when a photo is displayed.

For a video only stream with a single frame VideoOutput does not send an
EndOfVideoStream event.

VideoOutput doesn't immediately forward the EndOfVideoStream event to
MediaPlayer anymore. Moved all logic to decide when to send the
EndOfVideoStream event to VideoPlayer::startFrameTimer().