For photos "skip back" always skips to the previous play list entry
When displaying photos "skip back" now always skips to the previous file
in the play list. For video this is the case in the first second only.

The implementation counts the updates of the current time displayed. The
method "void SignalDispatcher::on_set_time(double seconds)" is called
twice when skipping to a photo.