VideoAttributeWidget is updated when another application changes Xv attributes
All X11/Xv dependencies are in XlibFacade.*pp.
All Gtk/Gdk dependencies are in gui directory.

GDK Window is used to add a filter function for events from the window.
This is in GtkmmMediPlayer / GtkmmAddGuiEventFilterFunctor.

The filter function itself is X11/Xv specific. It is implemented in
XFVideo / XvGuiEventFilterFunctor.

On the interface between GUI and XFVideo the 2 pure virtual base classes
AddGuiEventFilterFunctor and GuiEventFilterFunctor are used.

To avoid recursion the VideoAttributeWidget does not emit signalVideoAttributeChanged,
when it is updated by a NotificationVideoAttribute signal.