Avoid deadlock when receiving many XvPortNotifyEvent messages
With applications like gxvattr it is possible to modify Xv attributes of
the Xv port used by this player. Moving the sliders in gxvattr quickly
results in many events. A deadlock can occur in Glib::Dispatcher.

Glib::Dispatcher is using a pipe to trigger the GUI thread. Sending many
messages within VideoOutput or from VideoOutput to MediaPlayer results in
sending messages from the GUI thread to the GUI thread. The application
may completely fill the pipe and the write blocks. If it is blocking in
the GUI thread, then a deadlock occurred, since only the GUI thread is
reading from the pipe.

Solution: Writing to the pipe triggers calling the
GtkmmMediaPlayer::processEventQueue method. That method processes all
queued events. The fix is not to write into the pipe when a call to the
method is still pending.